requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
mysql-connector-python==8.3.0
twilio==8.10.0
pytest==8.1.1
requests==2.32.5